Possible fix for #112. Ensure the viewport is set each time a framebuffer is bound.
My understanding of gl.viewport is that it should be called each time the dimensions of the render area change. This
happens any time we render to a scaled picking target with a scale != 1.0. Not setting the viewport before rendering
was causing the picking pass to sometimes be rendered at full scale, which caused incorrect values to be returned by
the readpixels call as it was using the scaled x,y coordinates.

Since the overhead of gl.viewport is very small I think it's ok to call it as part of the framebuffer binding, just to be safe.